# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management SaaS API
  version: 1.0.0
  description: REST API for team projects, task assignment, JWT auth, RBAC, and task status tracking.
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://task-manager-saas.herokuapp.com/api
    description: Production PaaS URL placeholder
security:
  - bearerAuth: []
tags:
  - name: Authentication
  - name: Users
  - name: Projects
  - name: Tasks
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a user and return JWT tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate a user and return JWT tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token and return a new token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Token revoked.
        '400':
          $ref: '#/components/responses/ValidationError'
  /users:
    get:
      tags: [Users]
      summary: List users for assignment workflows. Admin only.
      responses:
        '200':
          description: Users returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects:
    get:
      tags: [Projects]
      summary: List projects visible to the current user.
      responses:
        '200':
          description: Projects returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Projects]
      summary: Create a project. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Project created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get one project.
      responses:
        '200':
          description: Project returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Projects]
      summary: Update a project. Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Projects]
      summary: Delete a project. Admin only.
      responses:
        '204':
          description: Project deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/members:
    post:
      tags: [Projects]
      summary: Assign a user to a project. Admin only.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Member assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/members/{userId}:
    delete:
      tags: [Projects]
      summary: Remove a project member. Admin only.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks in a project visible to the current user.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Tasks returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Tasks]
      summary: Create a task in a project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Tasks]
      summary: Get a task.
      responses:
        '200':
          description: Task returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Tasks]
      summary: Update a task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Tasks]
      summary: Delete a task.
      responses:
        '204':
          description: Task deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{taskId}/status:
    get:
      tags: [Tasks]
      summary: Get current task status for polling-based real-time tracking.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Current task status returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{taskId}/status/stream:
    get:
      tags: [Tasks]
      summary: Subscribe to task status updates using Server-Sent Events.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Event stream of task status updates.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user does not have permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        role:
          type: string
          enum: [ADMIN, MEMBER]
          default: MEMBER
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, MEMBER]
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        description:
          type: string
          maxLength: 1000
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        description:
          type: string
          maxLength: 1000
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectInput'
        - type: object
          required: [id, createdBy, members, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            createdBy:
              type: string
              format: uuid
            members:
              type: array
              items:
                $ref: '#/components/schemas/User'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
          default: TODO
        assigneeId:
          type: string
          format: uuid
          nullable: true
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        assigneeId:
          type: string
          format: uuid
          nullable: true
    Task:
      allOf:
        - $ref: '#/components/schemas/TaskInput'
        - type: object
          required: [id, projectId, status, createdBy, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            projectId:
              type: string
              format: uuid
            createdBy:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    TaskStatus:
      type: object
      required: [taskId, projectId, status, updatedAt, streamUrl]
      properties:
        taskId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        updatedAt:
          type: string
          format: date-time
        streamUrl:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string